pyproject.toml
[project]
name = "tinysignal"
version = "0.1.0"
description = "TinySignal is a self-hosted, lightweight app that clusters Hacker News posts and delivers AI-generated summaries straight to your Telegram"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "duckdb>=1.3.2",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.3",
    "requests>=2.32.5",
    "schedule>=1.2.2",
    "scikit-learn>=1.7.1",
    "tqdm>=4.67.1",
]
